meta_agent/requirements.txt
openai>=1.30
python-dotenv>=1.0
pyyaml>=6.0
cryptography>=41.0
